# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6472ef60c1fa850d0e0744c5
  title: schoolmanagementsystem
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/student/{id}":
    delete:
      x-skaffolder-id: 6472f513c1fa850d0e074578
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6472f126c1fa850d0e074556
      x-skaffolder-resource: student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6472f513c1fa850d0e074572
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6472f126c1fa850d0e074556
      x-skaffolder-resource: student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
    post:
      x-skaffolder-id: 6472f513c1fa850d0e074570
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6472f126c1fa850d0e074556
      x-skaffolder-resource: student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/student"
  "/student/findBy_courses/{key}":
    get:
      x-skaffolder-id: 6472f551c1fa850d0e074591
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 6472f126c1fa850d0e074556
      x-skaffolder-resource: student
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - student
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
...
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 6472f541c1fa850d0e074584
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6472f162c1fa850d0e07455c
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6472f540c1fa850d0e07457c
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6472f162c1fa850d0e07455c
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
    post:
      x-skaffolder-id: 6472f540c1fa850d0e07457e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6472f162c1fa850d0e07455c
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/teacher"
  "/teacher/findBy_courses/{key}":
    get:
      x-skaffolder-id: 6472f551c1fa850d0e07458f
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 6472f162c1fa850d0e07455c
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    student:
      x-skaffolder-id: 6472f126c1fa850d0e074556
      x-skaffolder-id-db: 6472ef60c1fa850d0e0744c6
      x-skaffolder-id-entity: 6472f125c1fa850d0e074552
      x-skaffolder-url: /student
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        dob:
          type: integer
          x-skaffolder-id-attr: 6472f125c1fa850d0e074555
          x-skaffolder-type: Date
        name:
          type: string
          x-skaffolder-id-attr: 6472f125c1fa850d0e074553
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 6472f125c1fa850d0e074554
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 6472f285c1fa850d0e07456d
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 6472f125c1fa850d0e074552
          x-skaffolder-ent2: 6472f17ec1fa850d0e07455f
      required:
        - _id
        - name
    teacher:
      x-skaffolder-id: 6472f162c1fa850d0e07455c
      x-skaffolder-id-db: 6472ef60c1fa850d0e0744c6
      x-skaffolder-id-entity: 6472f161c1fa850d0e074559
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 6472f161c1fa850d0e07455a
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 6472f161c1fa850d0e07455b
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 6472f23bc1fa850d0e07456c
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 6472f161c1fa850d0e074559
          x-skaffolder-ent2: 6472f17ec1fa850d0e07455f
      required:
        - _id
...